requirements.txt
openai
elevenlabs
ari
python-dotenv
requests
